more reorg, plus another refactor
