+++ Cargo.toml
@@ -1,8 +1,6 @@
 [workspace]
 
 members = [
-	"p1",
-	"p2",
+	"problems-rs",
 	"bio-lib",
-	"mendels1stlaw"
 ]
