--- p1/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "p1"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bio-lib = { path = "../bio-lib" }
