--- p1/rosalind_dna.txt
@@ -1,1 +1,0 @@
-ACATCGTGACGTAGTATCTTAAAAACGAATATTTATGCTATACGATTGATCCAGCATCCGATGTTGATTTATTCGAGAATGGGTGAAAGGGGGAGACCGACGGTCCCTGGCGGTCTGGGTTTGTAGGAACCAACTTTCTCGTTTTAGCATAGGCGAACGCAGGCGGTTCGGAACGCGTATCTTGACCCTGCGCACCGTACCCCCTTGAGCCAGTCTGTTATCGTGGAACACAGCTGCATTCCTAGCTCTAGAGCTAACCTTCTTACTCGTGCCTTTTGGACCAGTTGTCCTCCATCGTCCATGCTAGTCCCACTCAAGATAGGCGGTTCGAAAGCACCTTCTACCTGCTCCAGTCATTCTGTCGTGACCGCCAACGCAGCTGAAGAATATACATGCAATGTATGTTGGCTTTCTTCTTATGGAGCACGGCTATTCGTGTGGAATTTGTGAAGGTTTCTTGACTAGTCTCTGCTTTGTCATGTTAATGGATCCTCATTCTTCTTCCAAAGCAGCGAAAAAGATAGGTGGTAACATCCAGTCGGGGGAGCATTGGGCAATTCCTTGGGGTCTAGGTTACACTCGCGACCCAATACAAGCGAGAAGGCGTGCAACGTGGTGTCTATAGACAGGTATAGGCACCACGGTGTAAACCTGCAGCAGGCCGCGCGGTCGAAAATAAGATGTGGGCTTCATTCTCGCACGGCCATAGCTTGGCGTCTAGTTATACAAGATCCAGAGATATGGCGTTCGGTGCCGGCTACAGTTAGTGACTGGGCTTGGAGATGCCCATTATTAGCTGTATGCCGTTGTAACGCAAGTAATACAGTCAATTAGTTGAGCCGGCTTTCCGTTCTCGTCCTGTCGCGGACTATGATG
