+++ problems-rs/data/rosalind_dna.txt
@@ -1,0 +1,1 @@
+CGTCAGTAAACGCAGTGATCAGTTCCGTGGAAGTCCCTCCCCCTTTGGGGAAGGAGAATTAAGAGCAGGTTCTTACATTCAGGTAGATTTATTTGACTTGGCTTGGCTAGGTTAACTAGCTCCCCTTCCGAGCCTCGGCGACTATGTCCGTAATTAAGGGGCGGGAGTGCCAATCATGAGTGTTATGGATTCCTACGTATACCTCAAAATTTTATGGGGTTTCGAAAAGATTTATTCATGCTGGAAAGTTAGTTAATGGCACCTCAGAGACTAACCCTCAACCACGGTTAATCCTTCCGCCAGAAGACCCTAACGCCGTAGTAGTATCGAGATCCATGGTACAACCTGGGAATCATGAGGTTCGTGAAGTGGATGCTTCACTTACGTCGCTTTTCTAGCGCAGTAGGTCAAGTTTCGCTCTTACCCCCAGCCTGTCATGAGGCCTGTCGATGCCTGCAGGACTTCAGCTTGCAGAGGTTAAAATTAGATAGGCCCGGCAAGGCGCGGCGGATTTACCGACAAGCCTTGCGCTGAATTATTAGCCCTATGAAGTTACGGAAAGCCCTGGCAAACCAGGACAGCCGCGCCGTTTAGACAAAGCACTTGGACCCGTATTCGTTTTATTCACATACGTCCCGATCCTAGATTAGCGATGTTGAAATACTGTGTTGCTGGTACCAGCTTGCGCAAGGGGGAGTCTAATTATCCTCGACCAGCTCTCACGCAGCGGCGTCCCTTCGATAAGCGCGGTCGAGCCGTGCTCAATCCTACCACAATGGTAGCTGATTTCGTAATCACACAATCGAGGACATTCGGTGTATTTGTAGCAGGATATCTCAATAGATGAAGTCGAAA
