+++ problems-rs/data/rosalind_hamm.txt
