+++ problems-rs/data/rosalind_iprb.txt
