+++ problems-rs/data/rosalind_prot.txt
