+++ problems-rs/data/rosalind_revc.txt
