+++ problems-rs/data/rosalind_rna.txt
